:sparkles: (use the access feature without having to pay gas) checkAccess
Gas fees is a barrier to entry in this web3 space. Users and creators can create gated content without neccessity to pay eth gas fee. If Allow Subscription feature is enabled, then they
will be prompted to pay gas fee

BREAKING CHANGE: For Access implementing `Allow Subscription` the method to use is now ```checkAcccessWithSubscriptionEnabled``` and is only valid for initialized tokens.
```checkAccess``` method now works for without initializing the token. it accepts contractObj of AccessToken type and a user address.

Closes #14